+++ INTERX/gateway/rosetta/data/main.go
@@ -1,0 +1,11 @@
+package data
+
+import (
+	"github.com/gorilla/mux"
+	"github.com/grpc-ecosystem/grpc-gateway/runtime"
+)
+
+// RegisterRequest is a function to register requests.
+func RegisterRequest(router *mux.Router, gwCosmosmux *runtime.ServeMux, rpcAddr string) {
+	RegisterNetworkRoutes(router, gwCosmosmux, rpcAddr)
+}