+++ INTERX/scripts/protocgen.sh
@@ -15,4 +15,8 @@ for dir in $proto_dirs; do
 		--grpc-gateway_out=paths=source_relative:./proto-gen \
 		--openapiv2_out=third_party/OpenAPI/ \
 		$(find "${dir}" -maxdepth 1 -name '*.proto')
+
+# hack auto-generated proto -> remove omitempty tag
+# find ./proto-gen -path -prune -o -name '*.pb.go' | xargs -n1 -IX bash -c "sed -e 's/,omitempty//' X > X.tmp && mv X{.tmp,}"
+
 done