+++ x/gov/handler_test.go
@@ -1724,3 +1724,153 @@ func TestHandler_ProposalSetNetworkProperty(t *testing.T) {
 	iterator = app.CustomGovKeeper.GetActiveProposalsWithFinishedVotingEndTimeIterator(ctx, ctx.BlockTime())
 	require.True(t, iterator.Valid())
 }
+
+func TestHandler_CreateProposalCreateRole_Errors(t *testing.T) {
+	proposerAddr, err := sdk.AccAddressFromBech32("kira1alzyfq40zjsveat87jlg8jxetwqmr0a29sgd0f")
+	require.NoError(t, err)
+
+	tests := []struct {
+		name         string
+		msg          *types.MsgProposalCreateRole
+		preparePerms func(t *testing.T, app *simapp.SimApp, ctx sdk.Context)
+		expectedErr  error
+	}{
+		{
+			"Proposer does not have Perm",
+			types.NewMsgProposalCreateRole(
+				proposerAddr,
+				types.Role(1),
+				[]types.PermValue{},
+				[]types.PermValue{types.PermClaimValidator},
+			),
+			func(t *testing.T, app *simapp.SimApp, ctx sdk.Context) {},
+			errors.Wrap(types.ErrNotEnoughPermissions, types.PermCreateRoleProposal.String()),
+		},
+		{
+			"role already exist",
+			types.NewMsgProposalCreateRole(
+				proposerAddr,
+				types.Role(1),
+				[]types.PermValue{types.PermClaimCouncilor},
+				[]types.PermValue{},
+			),
+			func(t *testing.T, app *simapp.SimApp, ctx sdk.Context) {
+				proposerActor := types.NewDefaultActor(proposerAddr)
+				err := app.CustomGovKeeper.AddWhitelistPermission(
+					ctx,
+					proposerActor,
+					types.PermCreateRoleProposal,
+				)
+				require.NoError(t, err)
+
+				app.CustomGovKeeper.CreateRole(ctx, types.Role(1))
+			},
+			types.ErrRoleExist,
+		},
+		{
+			"permissions are empty",
+			types.NewMsgProposalCreateRole(
+				proposerAddr,
+				types.Role(1000),
+				[]types.PermValue{},
+				[]types.PermValue{},
+			),
+			func(t *testing.T, app *simapp.SimApp, ctx sdk.Context) {
+				proposerActor := types.NewDefaultActor(proposerAddr)
+				err := app.CustomGovKeeper.AddWhitelistPermission(
+					ctx,
+					proposerActor,
+					types.PermCreateRoleProposal,
+				)
+				require.NoError(t, err)
+			},
+			types.ErrEmptyPermissions,
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			app := simapp.Setup(false)
+			ctx := app.NewContext(false, tmproto.Header{})
+
+			tt.preparePerms(t, app, ctx)
+
+			handler := gov.NewHandler(app.CustomGovKeeper)
+			_, err := handler(ctx, tt.msg)
+			require.EqualError(t, err, tt.expectedErr.Error())
+		})
+	}
+}
+
+func TestHandler_ProposalCreateRole(t *testing.T) {
+	proposerAddr, err := sdk.AccAddressFromBech32("kira1alzyfq40zjsveat87jlg8jxetwqmr0a29sgd0f")
+	require.NoError(t, err)
+
+	app := simapp.Setup(false)
+	ctx := app.NewContext(false, tmproto.Header{
+		Time: time.Now(),
+	})
+
+	// Set proposer Permissions
+	proposerActor := types.NewDefaultActor(proposerAddr)
+	err2 := app.CustomGovKeeper.AddWhitelistPermission(ctx, proposerActor, types.PermCreateRoleProposal)
+	require.NoError(t, err2)
+
+	properties := app.CustomGovKeeper.GetNetworkProperties(ctx)
+	properties.ProposalEndTime = 10
+	app.CustomGovKeeper.SetNetworkProperties(ctx, properties)
+
+	handler := gov.NewHandler(app.CustomGovKeeper)
+	res, err := handler(
+		ctx,
+		types.NewMsgProposalCreateRole(
+			proposerAddr,
+			types.Role(1000),
+			[]types.PermValue{
+				types.PermClaimValidator,
+			},
+			[]types.PermValue{
+				types.PermChangeTxFee,
+			},
+		),
+	)
+	require.NoError(t, err)
+
+	expData, _ := proto.Marshal(&types.MsgProposalCreateRoleResponse{ProposalID: 1})
+	require.Equal(t, expData, res.Data)
+
+	savedProposal, found := app.CustomGovKeeper.GetProposal(ctx, 1)
+	require.True(t, found)
+
+	expectedSavedProposal, err := types.NewProposal(
+		1,
+		types.NewCreateRoleProposal(
+			types.Role(1000),
+			[]types.PermValue{
+				types.PermClaimValidator,
+			},
+			[]types.PermValue{
+				types.PermChangeTxFee,
+			},
+		),
+		ctx.BlockTime(),
+		ctx.BlockTime().Add(time.Second*time.Duration(properties.ProposalEndTime)),
+		ctx.BlockTime().Add(time.Second*time.Duration(properties.ProposalEnactmentTime)+time.Second*time.Duration(properties.ProposalEndTime)),
+	)
+	require.NoError(t, err)
+	require.Equal(t, expectedSavedProposal, savedProposal)
+
+	// Next proposal ID is increased.
+	id, err := app.CustomGovKeeper.GetNextProposalID(ctx)
+	require.NoError(t, err)
+	require.Equal(t, uint64(2), id)
+
+	// Is not on finished active proposals.
+	iterator := app.CustomGovKeeper.GetActiveProposalsWithFinishedVotingEndTimeIterator(ctx, ctx.BlockTime())
+	require.False(t, iterator.Valid())
+
+	ctx = ctx.WithBlockTime(ctx.BlockTime().Add(time.Minute * 10))
+	iterator = app.CustomGovKeeper.GetActiveProposalsWithFinishedVotingEndTimeIterator(ctx, ctx.BlockTime())
+	require.True(t, iterator.Valid())
+}