+++ INTERX/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+## [v0.1.x.x] - xx.xx.xxxx
+
+## [v0.1.17.5] - 03.17.2021
+
+### Updated
+
+- update cors allow headers
+
+### Added
+
+- add rosetta - network status api
+
+## [v0.1.17.4] - 03.16.2021
+
+### Updated
+
+- updated validators query to have `top` field - sort validators by `top` field
+- update interx metadata - sync to latest
+
+### Added
+
+- add rosetta - network list api
+- add rosetta - network options api
+
+## [v0.1.17.3] - 03.08.2021
+### Fixed
+- Validators query to include mischance.
+- Updated validators statistics
+- Updated validators query to include zero values.
+
+### Added
+- Tokens alias/rate query.
+- Voters/votes query.
+- Added configuration for https/http
+
 ## [v0.1.17.1] - 02.25.2021
 ### Added
 - Add proposals query